use member pair  in addition gen. model #32
